aes: a resource that does not fit the pool loads into far memory (far-trees step 2)
docs/far-trees.md, step 2 of 3.  QED's resource is 34 KB and the pool is
14 KB of bank $00; after step 1 the AES could DRAW a far tree but could
not LOAD one, because rsrc.c addressed a resource through a near RSHDR *
and fixed it up in place with a 16-bit base.

STAGE A -- one fixup, on any base.  A resource is now addressed through a
32-bit base, and one in the pool is a base whose bank is zero: the rule
step 1 gave trees.  The fixup reads each record into a scratch through
far_get, swaps and fixes it there, writes it back with far_put; the
header is copied out where a function needs its counts; rsrc_gaddr and
rsrc_saddr compute addresses rather than dereference pointers; the two
resource slots in the process record are 32-bit bases with a far mark
beside each pool mark.  Done FIRST, with no far resource in existence,
so the gates could judge the shared path on every resource this tree
already loads: m4, m7, m9, m12 (the file layer), m13 (icons -- the
image-bits-to-far path), m17 (the desktop, 320 frames after GO as
before), m22 and m23 (accessories with their own resources), m28 (an
accessory beside the desktop, two slots live at once).  All nine
identical.  LoRAM 264 free, Near 323, as before.

STAGE B -- the far path, and who may take it.  When the file will not fit
the pool and the caller has set bit 0 of int_in[0], rs_load takes ONE
far_alloc -- so the whole resource is inside a bank and tree[obj] is safe
to index -- streams the file up through a 128-byte buffer, and fixes it
up with the far base through the same code.  Otherwise it returns 0.
The opt-in cannot be set by accident: aes_entry zeroes int_in[] and
copies only control[1] words, and the small-data kit passes none.  The
large-data kit's rsrc_load passes one word set to 1 and keeps all 24 bits
of rsrc_gaddr's answer; the small-data kit keeps sixteen, which is all a
program of that model can hold.  A 16-bit program handed a far resource
would have had its pointers truncated silently, which is the failure
class this whole tree exists to refuse.  global[5..8] carry ap_ptree and
the header as 32-bit addresses, as the ST's do.  Colour icons stay a
pool-only feature.

STAGE C -- a resource that cannot fit, loaded both ways.  tools/farrsc.py
builds FARRSC.RSC: 42,364 bytes, 702 objects in 28 trees, three times the
pool and inside one bank.  src/m33_farrsc.c is built twice from one
source -- large-data as M33.G4A, small-data as M33S.G4A -- and
make test-m33 runs both under the stand-in desktop.  The large build got
its resource at $070000, tree 0 at $076338 inside it, drew the 40x28
dialog from there (12,400 dark pixels where an empty box has none),
walked 27 far objects with objc_find and landed on the OK button, read a
free string through the address it was handed, and gave the block back.
The small build was refused, and nothing else happened to it.

Two of the first gate's failures were its own: a program's near region
is the shell's again the instant it exits, so a result read after exit
is the desktop's variable; and sh_runs counts the desktop's return as a
run.  The program now waits at step 9 to be read, and the gate counts.

The oracle for the whole step, after m33: the full suite, 41 gates and
205 host tests, every one green (build/fartrees-suite2.log); then
make sdk, and GACS and RetroWP rebuilt against the rebuilt kit and run
end to end.  Both pass -- neither asks for a far resource, and their
resources load into the pool through the one fixup that now serves both.

+++ src/aes/fsel.c
@@ -547,7 +547,7 @@ WORD fs_input(char *pipath, char *pisel, WORD *pbutton, const char *pilabel)
     mask = locold + LEN_FSPATH;
 
     far_get((uint8_t *)h, (uint32_t)(const uint8_t FAR *)fs_rsc, FS_RSC_SIZE);
-    rs_fixit(h);
+    rs_fixit((uint32_t)(uint16_t)h);   /* a bank-zero base: the pool */
     tree = (OBJECT FAR *)(uint16_t)*(uint32_t *)((uint8_t *)h + h->rsh_trindex);
 
     /* the donor's fs_start: centred, and the scroll bar the width of a
